backend/pyproject.toml
[tool.poetry]
name = "pyproject"
version = "0.1.0"
description = ""
authors = ["hank <h31169638@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
celery = "^5.3.6"
django = "^5.0.4"
django-cors-headers = "^4.3.1"
django-redis = "^5.4.0"
djangorestframework = "^3.15.1"
django-ratelimit = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
